+++ src/Drivers/Basic/BasicExtensionDriver.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace wapmorgan\UnifiedArchive\Drivers\Basic;
+
+abstract class BasicExtensionDriver extends BasicDriver
+{
+
+}