+++ src/Drivers/Basic/BasicUtilityDriver.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace wapmorgan\UnifiedArchive\Drivers\Basic;
+
+class BasicUtilityDriver extends BasicDriver
+{
+}