+++ src/Drivers/SplitbrainPhpArchive.php
@@ -1,0 +1,113 @@
+<?php
+
+namespace wapmorgan\UnifiedArchive\Drivers;
+
+use wapmorgan\UnifiedArchive\Drivers\Basic\BasicDriver;
+
+class SplitbrainPhpArchive extends BasicDriver
+{
+
+    /**
+     * @inheritDoc
+     */
+    public static function getDescription()
+    {
+        // TODO: Implement getDescription() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public static function isInstalled()
+    {
+        // TODO: Implement isInstalled() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public static function getInstallationInstruction()
+    {
+        // TODO: Implement getInstallationInstruction() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public static function getSupportedFormats()
+    {
+        // TODO: Implement getSupportedFormats() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public static function checkFormatSupport($format)
+    {
+        // TODO: Implement checkFormatSupport() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function getArchiveInformation()
+    {
+        // TODO: Implement getArchiveInformation() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function getFileNames()
+    {
+        // TODO: Implement getFileNames() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function isFileExists($fileName)
+    {
+        // TODO: Implement isFileExists() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function getFileData($fileName)
+    {
+        // TODO: Implement getFileData() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function getFileContent($fileName)
+    {
+        // TODO: Implement getFileContent() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function getFileStream($fileName)
+    {
+        // TODO: Implement getFileStream() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function extractFiles($outputFolder, array $files)
+    {
+        // TODO: Implement extractFiles() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function extractArchive($outputFolder)
+    {
+        // TODO: Implement extractArchive() method.
+    }
+}