+++ tests/ArchivingTest.php
@@ -27,7 +27,7 @@ class ArchivingTest extends PhpUnitTestCase
 
         $test_archive_filename = WORK_DIR.'/'.$archiveFileName;
 
-        $result = UnifiedArchive::archive(FIXTURES_DIR, $test_archive_filename);
+        $result = UnifiedArchive::create(FIXTURES_DIR, $test_archive_filename);
         $this->assertValueIsInteger($result);
         $this->assertEquals(5, $result);
 
@@ -54,7 +54,7 @@ class ArchivingTest extends PhpUnitTestCase
 
         $test_archive_filename = WORK_DIR.'/'.$archiveFileName;
 
-        $result = UnifiedArchive::archive(FIXTURES_DIR . '/doc', $test_archive_filename);
+        $result = UnifiedArchive::create(FIXTURES_DIR . '/doc', $test_archive_filename);
         $this->assertValueIsInteger($result);
         $this->assertEquals(1, $result);